Se logra obtener acciones
El bot es capaz de obtener acciones de la bolsa en tiempo real.
