+++ backend/main/java/edu/unlam/asistente/asistente_virtual/Bot.java
@@ -10,7 +10,8 @@ import edu.unlam.asistente.chucknorris.ChuckNorrisFact;
 import edu.unlam.asistente.clima.DiaLluvioso;
 import edu.unlam.asistente.conversor_unidades.ConversorUnidad;
 import edu.unlam.asistente.cordialidad.Cordialidad;
-import edu.unlam.asistente.financiera.ValorMoneda;
+import edu.unlam.asistente.financiera.acciones.Acciones;
+import edu.unlam.asistente.financiera.moneda.ValorMoneda;
 import edu.unlam.asistente.imagenes.gif.Gif;
 import edu.unlam.asistente.leyes_robotica.LeyesRobotica;
 import edu.unlam.asistente.noticias.Noticias;
@@ -52,6 +53,7 @@ public class Bot implements IDecision {
 		DiaLluvioso diaLluvioso = new DiaLluvioso();
 		Noticias noticias = new Noticias();
 		Gif gif = new Gif();
+		Acciones acciones = new Acciones();
 		ValorMoneda valorMoneda = new ValorMoneda();
 		Default def = new Default();
 
@@ -65,7 +67,8 @@ public class Bot implements IDecision {
 		busquedaWeb.setSiguienteDecision(gif);
 		gif.setSiguienteDecision(diaLluvioso);
 		diaLluvioso.setSiguienteDecision(noticias);
-		noticias.setSiguienteDecision(valorMoneda);
+		noticias.setSiguienteDecision(acciones);
+		acciones.setSiguienteDecision(valorMoneda);
 		valorMoneda.setSiguienteDecision(calendario);
 
 		// @def debe quedar siempre al final dado que es el que mostrará el
