+++ backend/main/java/edu/unlam/asistente/financiera/moneda/Rublo.java
@@ -1,15 +1,15 @@
-package edu.unlam.asistente.financiera;
+package edu.unlam.asistente.financiera.moneda;
 
 import java.util.regex.Pattern;
 
 /**
- * Clase que administra la cotización del peso Chileno. <br>
+ * Clase que administra la cotización del Rublo. <br>
  */
-public class Chileno implements Cotizacion {
+public class Rublo implements Cotizacion {
 	/**
-	 * Regex del peso Chileno. <br>
+	 * Regex del Rublo. <br>
 	 */
-	private static final Pattern REGEX_CHILENO = Pattern.compile("(?:chileno|Chileno)");
+	private static final Pattern REGEX_RUBLO = Pattern.compile("(?:rublo|Rublo)");
 	/**
 	 * Indica cuál es la siguiente cotización que debe intentar resolver la
 	 * solicitud. <br>
@@ -18,8 +18,8 @@ public class Chileno implements Cotizacion {
 
 	@Override
 	public String leerMoneda(String moneda) {
-		if (REGEX_CHILENO.matcher(moneda).find()) {
-			return Moneda.obtenerCotizacionActual("CLP");
+		if (REGEX_RUBLO.matcher(moneda).find()) {
+			return Moneda.obtenerCotizacionActual("RUB");
 		}
 		return this.siguienteCotizacion.leerMoneda(moneda);
 	}
