+++ backend/main/java/edu/unlam/asistente/financiera/moneda/Uruguayo.java
@@ -1,4 +1,4 @@
-package edu.unlam.asistente.financiera;
+package edu.unlam.asistente.financiera.moneda;
 
 import java.util.regex.Pattern;
