+++ backend/main/java/edu/unlam/asistente/financiera/moneda/ValorMoneda.java
@@ -1,10 +1,13 @@
-package edu.unlam.asistente.financiera;
+package edu.unlam.asistente.financiera.moneda;
 
 import java.util.regex.Matcher;
 import java.util.regex.Pattern;
 
 import edu.unlam.asistente.asistente_virtual.IDecision;
 
+/**
+ * Clase para obtener el valor actual de alguna moneda solicitada. <br>
+ */
 public class ValorMoneda implements IDecision {
 	/**
 	 * Regex del valor de alguna moneda. <br>
@@ -21,7 +24,9 @@ public class ValorMoneda implements IDecision {
 	public String leerMensaje(String mensaje, String usuario) {
 		Matcher matcher = REGEX_VALOR_MONEDA.matcher(mensaje);
 		if (matcher.find()) {
-			return new Moneda().leerMoneda(matcher.group());
+			String moneda = matcher.group();
+			return new StringBuilder("@").append(usuario).append(" su valor es de ")
+					.append(new Moneda().leerMoneda(moneda)).toString();
 		}
 		return this.siguienteDecision.leerMensaje(mensaje, usuario);
 	}
