+++ backend/tests/edu/unlam/asistente/asistente_virtual/BotTests.java
@@ -636,6 +636,18 @@ public class BotTests {
 	
 	@Test
 	public void obtenerValorDolar() {
-		Assert.assertTrue(Double.parseDouble(jenkins.leerMensaje("@jenkins, cual es el valor del dolar", USUARIO)) > 25.0);
+		Assert.assertTrue(Double.parseDouble(jenkins.leerMensaje("@jenkins, cual es el valor del dolar", USUARIO)
+				.substring(USUARIO.length() + 18)) > 25.0);
+	}
+
+	@Test
+	public void obtenerAcciones() {
+		String datos = jenkins.leerMensaje("@jenkins, como fueron las acciones de GOOGL?", USUARIO);
+		Assert.assertNotEquals(
+				"@" + USUARIO
+						+ " el nombre de la empresa (BCBA) no esta bien escrito, por favor escribalo correctamente.",
+				datos);
+		Assert.assertNotEquals(
+				"@" + USUARIO + " ha ocurrido un problema al obtener los datos financiero. Intente más tarde.", datos);
 	}
 }
