+++ backend/tests/edu/unlam/asistente/financiera/moneda/MonedaTest.java
@@ -1,4 +1,4 @@
-package edu.unlam.asistente.financiera;
+package edu.unlam.asistente.financiera.moneda;
 
 import java.util.Arrays;
 import java.util.Collection;
